Introduce support for optimistic parsing
This commit adds "optimistic" tokenizers and parsers to the `plainjs`
codebase. Optimistic tokenizers will emit "incomplete" tokens when they
think that they will eventually reach a state where a token can be
definitively produced. Optimistic parsers respond to incomplete tokens
and other states by presenting a view of the parsed value that is
expected to be in some sense a "subset" of the value that will
eventually be parsed (assuming, optimistically, that the stream will
constitute a valid value when all is said and done).